DucAnh/SignLanguage_Data/requirements.txt
mediapipe==0.10.14
numpy==2.0.2
pyserial==3.5
opencv-python==4.10.0.84
tensorflow==2.19.0
keras==3.10.0 
pandas==2.2.2